Need to get that column as a byte so that it is boxed into a Byte otherwise we'll end up with an Integer which reflection doesn't convert to a Byte when coercing the method parameters.
git-svn-id: https://samskivert.googlecode.com/svn/trunk@1981 6335cc39-0255-0410-8fd6-9bcaacd3b74c

+++ src/java/com/samskivert/jdbc/depot/FieldMarshaller.java
@@ -420,7 +420,7 @@ public abstract class FieldMarshaller
         public void getValue (ResultSet rs, Object po)
             throws SQLException, IllegalAccessException {
             try {
-                _field.set(po, _factmeth.invoke(null, rs.getInt(getColumnName())));
+                _field.set(po, _factmeth.invoke(null, rs.getByte(getColumnName())));
             } catch (InvocationTargetException ite) {
                 throw new RuntimeException(ite);
             }
